backend/requirements.txt
Django
djangorestframework
pymongo
djongo
python-dotenv